backend/requirements.txt
# Python 3.13 compatible dependency set
# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# AI/LLM
openai>=1.55.0
anthropic>=0.40.0

# Vector DB
pinecone>=6.0.0

# Data / config
pydantic>=2.10.0
python-dotenv>=1.0.1

# HTTP / testing
httpx>=0.27.0
pytest>=8.3.0
